+++ statham/dsl/elements/array.py
@@ -1,6 +1,6 @@
 from typing import List, TypeVar
 
-from statham import validators as val
+from statham.dsl import validators as val
 from statham.dsl.elements.base import Element
 from statham.dsl.constants import Maybe, NotPassed
