+++ statham/dsl/elements/string.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from statham import validators as val
+from statham.dsl import validators as val
 from statham.dsl.elements.base import Element
 from statham.dsl.constants import Maybe, NotPassed
