+++ tests/component/test_array_item_validation.py
@@ -1,6 +1,6 @@
 import pytest
 
-from statham.exceptions import ValidationError
+from statham.dsl.exceptions import ValidationError
 from tests.helpers import no_raise
 from tests.models.array_item_validation import Model
